% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normpatch-package.R
\docType{package}
\name{normpatch-package}
\alias{normpatch}
\alias{normpatch-package}
\title{normpatch}
\description{
Robust normalization factor estimation for RNA-seq count data
}
\details{
For a comprehensive introduction to the package,
including a quick start and case studies,
see the package documentation site:
\url{https://bitdessin.github.io/normpatch}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bitdessin/normpatch}
  \item Report bugs at \url{https://github.com/bitdessin/normpatch/issues}
}

}
\author{
\strong{Maintainer}: Jianqiang Sun \email{sun@bitdessin.dev} (\href{https://orcid.org/0000-0002-3438-3199}{ORCID})

}
\keyword{internal}
